returned rid and sid values back to int's from byte's, and fixed the …

+++ header.h
@@ -15,15 +15,15 @@
 typedef struct header Header;
 
 struct header {
-	BYTE sid;					// Sender ID
-	BYTE rid;					// Receiver ID
-	BYTE rid2;					// Used for voteOn() error detection and correction
-	BYTE rid3;					// Used for voteOn() error detection and correction
+	int sid;					// Sender ID
+	int rid;					// Receiver ID
+	int rid2;					// Used for voteOn() error detection and correction
+	int rid3;					// Used for voteOn() error detection and correction
 	BYTE priority;				// Message's pirority
 	int seqNum;					// Message's position in the queue
 	long payloadSize;			// Number of bytes in payload after this header
 	BYTE payloadType;			// mERR, mNONE, mTXT, mAUD, numMsgTypes
-	BYTE encryption;				// ERR, NONE, XOR, VIG, numOfEnc
+	BYTE encryption;			// ERR, NONE, XOR, VIG, numOfEnc
 	BYTE compression;			// cERR, cNONE, cHUF, cRLE, numCompTypes
 	long uncompressedLength;	// Uncompressed message length
 	int checksum;				// Checksum for payload error detection
